added jwt auth api.

+++ controllers/api/v1/post_api.js
@@ -21,13 +21,25 @@ module.exports.index = async function(req,res){
 
 module.exports.destroy = async function(req,res){
     try{
-        await Post.findByIdAndDelete(req.params.id);
-        await  Comment.deleteMany({post:req.params.id})
 
-        return res.status(200).json({
-            message:'Post and associated comments deleted successfully!'
-        })
+        const post = await Post.findById(req.params.id);
+
+        if(String(post.user) == req.user._id){
+            post.deleteOne();
+
+            await  Comment.deleteMany({post:req.params.id})
+
+            return res.status(200).json({
+                message:'Post and associated comments deleted successfully!'
+            });
+        }else{
+            return res.status(401).json({
+                message:'You cannot delete this post!'
+            });
+        }
+        
     }catch(err){
+        console.log(err);
         return res.status(500).json({
             message:'Internal server error'
         })
